SpreadEdge/.history/backend/requirements_20250417121709.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==1.0.3
stripe==7.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
requests==2.31.0
python-binance==1.0.19
ccxt==4.1.13
pandas==2.1.2
numpy==1.24.3
scikit-learn==1.3.2 